# Optimization: use RBTree's iterator.update in put and batch