@nolanlawson 2016-01-24
Build is failing because Saucelabs only runs for contributors... it's a dumb limitation, and I'll get around it by just opening a PR myself with your exact commits. Once that's green, I'm +1.