@nolanlawson 2016-01-24
https://github.com/Level/memdown/pull/45